Call paramin

Call paramin

callParamin(
  i = "params.in",
  func = "gadget -s -n",
  opt = "optinfo",
  network = "network",
  o = "params.out",
  scale = NULL,
  condor = NULL,
  paramin.exe = "paramin"
)

Arguments

i

input file

func

function name

opt

location of the optinfofile

network

location of the network file

o

output filename

scale

deprecated

condor

deprecated

paramin.exe

location of the paramin binary

Value

null

Details

Paramin runs the optimisation routines in parallel across a network of processors which will (hopefully) result in a much faster optimisation run. The current version of paramin is implemented with MPI, Message Passing Interface, which handles all the message passing between the master process, paramin, and the slave processes, which are Gadget simulation runs. The setup is very similar to a normal Gadget run, with all the Gadget input files are the same, it's only the optimisation execution that differs.

Author

Bjarki Thor Elvarsson