http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:surveyindices

http://hafro.github.io/gadget/docs/userguide/chap-like.html#sec:surveyindices

gadget_surveyindices_component(
  weight = 0,
  name = "surveyindices",
  sitype = "lengths",
  biomass = 0,
  data = NULL,
  area = NULL,
  fittype = NULL,
  length = NULL,
  age = NULL,
  fleetnames = NULL,
  surveynames = NULL,
  stocknames = NULL,
  slope = NULL,
  intercept = NULL
)

Arguments

weight

A numeric weighting

name

A descriptive name for the component, default same as component type

sitype

Survey index type. One of 'lengths'/'ages'/'fleets'/'acoustic'/'effort'

biomass

Whether the index data should be based on the biomass of the stock or on the population numbers for the stock. 0 or 1

data

data.frame required by the data_function

area

Area aggregation, if not attached to the data.frame by MFDB

fittype

Type of linear regression to use. One of 'linearfit' / 'loglinearfit' / 'fixedslopelinearfit' / 'fixedslopeloglinearfit' / 'fixedinterceptlinearfit' / 'fixedinterceptloglinearfit' / 'fixedlinearfit' / 'fixedloglinearfit'.

length

Length aggregation, if not attached to the data.frame by MFDB

age

Age aggregation, if not attached to the data.frame by MFDB

fleetnames

A list of all the fleets to be aggregated into a single pseudo fleet for the purposes of the data comparison

surveynames

A list of all the stocks to be aggregated into a single pseudo stock

stocknames

A list of all the stocks to be aggregated into a single pseudo stock

slope

Parameter for linear regression

intercept

Parameter for linear regression