Gadget fit

gadget.fit calls Gadget and compiles fit statistics, relevant biomass trajectories and consumption. It returns an object of class 'gadget.fit' which is essentially a list containing the following objects:

sidat

data.frame containing the model fit to the surveyindices

resTable

Table of likelihood component scores from the different stages of the iterative reweighting run

nesTable

Same as resTable but normalised with the minimum score for each component

suitability

The model estimated fleet predator suitability

stock.recruitment

Recruitment by stock

res.by.year

Results by year

stomachconten

Fit combined with data from the stomachcontent likelihood

likelihoodsummary

Likelihood component scores by year and step

catchdist.fleets

Data compared with model output for the catchdistribution components

stockdist

Data compared with model output for the stockdistribution components

SS

The likelihood output from the model

stock.full

Modeled abundance and mean weight by year,step, length and stock

stock.std

Modeled abundance, mean weight, number consumed by age, stock and year

stock.prey

Consumption of prey (or fleet catches) by prey, length, year and step

fleet.info

Information on catches, harvest rate and harvestable biomass by fleet, year and step

predator.prey

consumption of preys by predators, length, year and step

params

parameter values used for the fit

catchstatistics

Data compared with model output for the catchdstatistics components

bind.gadget.fit combines two (or more) gadget.fit objects into a single list. Each object in the list will have an extra column indicating the model, if the list is named those names will be used to identify the model.

gadget.fit(
  wgts = "WGTS",
  main.file = NULL,
  fleet.predict = NULL,
  mat.par = NULL,
  params.file = NULL,
  f.age.range = NULL,
  fit.folder = "FIT",
  printatstart = 1,
  steps = 1,
  recruitment_step_age = NULL,
  gd = NULL
)

bind.gadget.fit(...)

Arguments

wgts
Location of the iterative reweighting folder, if null gadget.fit requires a parameter file

main.file
if the main file is different from the 'main' or 'wgts/main.final'

fleet.predict
data.frame containing the basis fleets used to calculate the harvestable biomass. Fleet names should be specified in the fleet column.

mat.par
parameters for the maturity ogive

params.file
parameter file used in the fit (defaults to "WGTS/params.final")

f.age.range
data.frame describing the desired age range where the F's are calculated, if null this defaults to the apical F fro all stocks. Input columns should include stock, age.min and age.max

fit.folder
location of the output

printatstart
should the stock standard output be printed at the beginning or the end of the timestep

steps
what steps should be printed

recruitment_step_age
data frame defining the recruitment age by stock and time step. Default is the minimum age at step 1. Expects columns stock, age and step

gd
gadget_directory object to optionally set the location of the gadget directory

...
any number of gadget.fit objects

Value

list containing the output from Gadget.